Fixed some bugs where the user could go 'back' to see the content of authorized only pages. Also made a better back button so the user will be taken to a higher url domain insted of just back to where they came from
