+++ src/app/stian/brygging/home/page.tsx
@@ -1,19 +1,8 @@
 'use client'
-import { use, useEffect } from 'react';
-import { useAuth } from '../lib/authentication';
+import { useEffect } from 'react';
 
 export default function Home() {
 
-    const { logOut } = useAuth();
-
-    const handleLogOut = async () => {
-        try {
-            await logOut();
-
-        } catch (error) {
-            console.error(error);
-        }
-    }
 
     useEffect(() => {
         console.log("Home page")
@@ -22,7 +11,6 @@ export default function Home() {
 
     return (
         <div>
-            <button onClick={() => {handleLogOut}}>Logout</button>
             Welcome to the Brygging app!
         </div>
     )
