+++ src/app/stian/brygging/home/profile/page.tsx
@@ -1,0 +1,7 @@
+export default function Profile() {
+    return (
+        <div>
+            Profile
+        </div>
+    )
+}
