--- src/app/stian/brygging/lib/navbar.tsx
@@ -1,20 +1,0 @@
-export default function Navbar() {
-    return (
-        <div>
-            <ul>
-                <li>
-                    Home
-                </li>
-                <li>
-                    About
-                </li>
-            </ul>
-
-            <ul>
-                <li>
-                    Login
-                </li>
-            </ul>
-        </div>
-    );
-}
