+++ src/app/stian/ui/back-button.tsx
@@ -1,10 +1,20 @@
-import { useRouter } from 'next/navigation';
+import { useRouter, usePathname } from 'next/navigation';
 import "./back-button.css"
 export default function BackButton() {
     const router = useRouter();
+    const pathName = usePathname();
+
+    const goBackOnePage = () => {
+        const pathSegments = pathName.split('/').slice(0, -1);
+        const newPath = pathSegments.length > 1 ? pathSegments.join('/') : '/';
+        console.log(pathSegments.length);
+        console.log(newPath);
+        router.push(newPath);
+    }
+    
     return (
         <div>
-            <button className="back-button" type="button" onClick={() => router.back()}>
+            <button className="back-button" type="button" onClick={goBackOnePage}>
                 Click here to go back
             </button>
         </div>
